BETA 3 commit
This is not the BETA 3 announced before. I decided
to name this beta 3 to avoid problems to release so a big
change as beta 2, the second release.
The new beta number will allow me to create new packages
which are needed after all that changes.

This is commit will fix some small problems and
- new wilderness maps (rusty)
- new underground dungeon (cat)
- dark temple map (NOT COMMITED HERE!)
- dozen new arches
- dozens new monsters
- reworked exp/level/mob stat system
- rework color target table
- alot more small fixes

+++ server/src/include/libproto.h
@@ -54,7 +54,7 @@ extern void player_lvl_adj(object *who, object *op);
 extern void calc_perm_exp(object *op);
 extern int adjust_exp(object *pl, object *op, int exp);
 extern void apply_death_exp_penalty(object *op);
-extern int calc_level_difference(int who_lvl, int op_lvl);
+extern float calc_level_difference(int who_lvl, int op_lvl);
 /* friend.c */
 extern void add_friendly_object(object *op);
 extern void remove_friendly_object(object *op);
